requirements.txt
Flask==1.1.2
Flask-RESTful==0.3.6
flask_restful_swagger==0.20.2
flask_pymongo
pyjwt
flask_cors